pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taipy-chat-demo"
version = "0.1.0"
description = "Demo apps including a Taipy chat GUI and YouTube downloader."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
# Runtime dependencies
dependencies = [
    "taipy",
    "openai",
    "python-dotenv>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
